requirements.txt
Django==4.0.5
python-decouple==3.6
djangorestframework==3.13.1
psycopg2-binary==2.9.3
dj-database-url==0.5.0
django-rest-registration==0.7.0
djangorestframework-simplejwt==5.0.0

